+++ src/components/Keyboard/index.js
@@ -1,5 +1,0 @@
-import ConverterKeyboard from './ConverterKeyboard';
-
-import './Keyboard.css';
-
-export { ConverterKeyboard };