+++ src/components/Lists/index.js
@@ -1,0 +1,2 @@
+export { default as MenuList } from './Menu';
+export { default as ContentList } from './Content';