+++ src/components/Menu/index.js
@@ -1,12 +1,13 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
 
-import MenuList from '../MenuList';
+import { MenuList } from '../Lists';
 
 import './Menu.scss';
 
 const Menu = () => (
   <nav className="menu">
+    <h2>Converters</h2>
     <ul>
       <li>
         <NavLink to="/currency">Currency</NavLink>