+++ src/components/Title/index.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
 import './Title.scss';
 
@@ -7,5 +8,9 @@ import { toTitle, locationToTitle } from '../../helpers';
 const Title = ({ location }) => (
   <h2 className="screen__title">{locationToTitle(toTitle(location))}</h2>
 );
+
+Title.propTypes = {
+  location: PropTypes.string.isRequired
+};
 
 export default Title;